Included an example in README.md
